Added firebase database configuration
None of the models nor the controllers have been defined as of yet.
But, firebase has connected to the project, and this has taken too
much of my time.

Next up, the controllers and models will be defined, I'll think about
how to connect these to firebase at a later stage.