Database, routes, and failed controllers
The database has been connected to the application.
Routing has been made, but, isn't functional due to errors to do with
the definition of the controllers.
Controllers throw an error mostly to do with improper configuration of
the responses.

The more I try to get this to start working, the more I gravitate
towards having MongoDB as the database.
The reason why?

- Firebase is more of a backend, in and of itself. This means that there
  is no specific feature that needs to be set up by the backend.
- Routing and the project creation can be simply done by adding
  everything to the front end instead.
- The documentation is actually really *really* confusing.